Add GitHub Actions workflow for publishing
This commit introduces a new GitHub Actions workflow named "Build" (publish.yaml) that automates the build and publishing process.

The workflow is triggered on pushes to the "master" branch or can be manually dispatched. It runs on an Ubuntu environment and performs the following steps:
- Checks out the code
- Sets up JDK 21
- Sets up Gradle, enabling dependency graph generation and build scan publishing
- Builds the project using Gradle and executes the `publish` task. It also configures Gradle JVM arguments for improved performance.

Additionally, the existing "Build" workflow (build.yml) is updated to skip execution if the commit message starts with "[skip ci]" and the repository owner is "RikkaApps".